Move the watchActiveLocation helper to a useActiveLocation composable
Also renamed the useCurrentRoute composable to useRoute to stay
consistent with naming and refactored the useRouteName composable to use
the new useRoute composable internally, which reduces code complexity
and actually makes it reactive.
