+++ packages/web-app-files/src/composables/index.ts
@@ -1,5 +1,6 @@
 export * from './fileListHeaderPosition'
 export * from './pagination'
+export * from './router'
 export * from './sort'
 
 declare module 'vue/types/vue' {
