+++ packages/web-app-files/src/router/index.ts
@@ -23,7 +23,7 @@ import {
   isLocationCommonActive,
   createLocationCommon
 } from './common'
-import { isAuthenticatedRoute, watchActiveLocation } from './utils'
+import { isAuthenticatedRoute, ActiveRouteDirectorFunc } from './utils'
 import { buildRoutes as buildDeprecatedRoutes, isLocationActive } from './deprecated'
 import { RouteComponents } from './router'
 import { RouteConfig } from 'vue-router'
@@ -56,6 +56,6 @@ export {
   isLocationPublicActive,
   isLocationActive,
   isAuthenticatedRoute,
-  watchActiveLocation,
-  buildRoutes
+  buildRoutes,
+  ActiveRouteDirectorFunc
 }
