+++ packages/web-pkg/src/composables/router/useRoute.ts
@@ -2,9 +2,10 @@ import { ref, Ref, readonly } from '@vue/composition-api'
 import { useRouter } from './useRouter'
 import { Route } from 'vue-router'
 
-export const useCurrentRoute = (): Ref<Route> => {
+export const useRoute = (): Ref<Route> => {
   const router = useRouter()
   const currentRoute = ref()
+  currentRoute.value = router.currentRoute
   router.afterEach((to) => (currentRoute.value = { ...to }))
 
   return readonly(currentRoute)
